Changed parameter to a float to remove compiler warning.

+++ deformable_manipulation_experiment_params/include/deformable_manipulation_experiment_params/ros_params.hpp
@@ -620,7 +620,7 @@ namespace smmap
         return ROSHelpers::GetParam(nh, "num_simsteps_per_gripper_command", 4);
     }
 
-    inline double GetSettlingTime(ros::NodeHandle& nh, const double default_time = 4.0)
+    inline float GetSettlingTime(ros::NodeHandle& nh, const float default_time = 4.0)
     {
         return ROSHelpers::GetParam(nh, "settle_time", default_time);
     }
